Move the default hasher one directory up (#2580)
The intermediate directory/package is empty, so move rfc6962/hasher to rfc6962.
It also used to be there previously, so this change brings backward compatibility for this
specific package back.
