+++ client/log_verifier_test.go
@@ -18,7 +18,7 @@ import (
 	"testing"
 
 	"github.com/google/trillian"
-	rfc6962 "github.com/google/trillian/merkle/rfc6962/hasher"
+	"github.com/google/trillian/merkle/rfc6962"
 	"github.com/google/trillian/types"
 )
