+++ merkle/rfc6962/rfc6962.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package hasher provides hashing functionality according to RFC6962.
-package hasher
+// Package rfc6962 provides hashing functionality according to RFC6962.
+package rfc6962
 
 import (
 	"crypto"
