+++ merkle/rfc6962/rfc6962_test.go
@@ -11,7 +11,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package hasher
+
+package rfc6962
 
 import (
 	"bytes"
